Work on RetroShell

+++ Emulator/RetroShell/Command.cpp
@@ -62,7 +62,7 @@ Command::add(const std::vector<string> &tokens,
     Command d;
     d.name = tokens.back();
     d.fullName = (cmd->fullName.empty() ? "" : cmd->fullName + " ") + tokens.back();
-    d.group = groups.size() - 1;
+    d.group = isize(groups.size()) - 1;
     d.requiredArgs = requiredArgs;
     d.optionalArgs = optionalArgs;
     d.help = help;
